add session and author

+++ show-article.php
@@ -1,6 +1,11 @@
 <?php
 
 require_once __DIR__ . '/database/database.php';
+
+require_once __DIR__ . '/database/security.php';
+$currentUser = isloggedin();
+
+
 $articleDB = require_once './database/models/ArticleDB.php';
 
 
@@ -26,7 +31,7 @@ if (!$id) {
 <html lang="en">
 
 <head>
-    <?php require_once 'includes/head.php' ?>
+    <?php require_once __DIR__ . '/includes/head.php' ?>
     <link rel="stylesheet" href="public/css/show-article.css">
     <title>Article</title>
 </head>
@@ -42,12 +47,18 @@ if (!$id) {
                 <h1 class="article-title"> <?= $article['title'] ?></h1>
                 <div class="separateur"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <!-- Mettez l'URL de suppression directement dans l'attribut "href" du lien -->
-                    <a class="btn btn-secondary" id="deleteButton" data-id="<?= $article['id'] ?>" href="/delete-article.php?id=<?= $article['id'] ?>" type="button">Supprimé</a>
-
-                    <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                <div class="article-author">
+                    <p><?= $article['firstname'] . ' ' . $article['lastname'] ?></p>
                 </div>
+                <?php if ($currentUser && $currentUser['id'] === $article['author']) : ?>
+                    <div class="action">
+                      
+                        <a class="btn btn-secondary" id="deleteButton" data-id="<?= $article['id'] ?>" href="/delete-article.php?id=<?= $article['id'] ?>" type="button">Supprimé</a>
+
+                        <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                    </div>
+                <?php endif; ?>
+
             </div>
 
         </div>
@@ -55,8 +66,12 @@ if (!$id) {
     </div>
 
 
+
     <div id="deleteModal" class="delete-modal">
         <div class="modal-content">
+            <div class="delete-message">
+                <img src="./public/img/deleteMessage.svg" alt="">
+            </div>
             <h2>Confirmation</h2>
             <p>Êtes-vous sûr de vouloir supprimer votre article?</p>
             <button id="confirmDelete">Confirmer</button>
@@ -66,28 +81,6 @@ if (!$id) {
 
 
 
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            const deleteButton = document.getElementById('deleteButton');
-            const deleteModal = document.getElementById('deleteModal');
-            const confirmDelete = document.getElementById('confirmDelete');
-            const cancelDelete = document.getElementById('cancelDelete');
-
-            deleteButton.addEventListener('click', function(event) {
-                event.preventDefault();
-                deleteModal.style.display = 'block';
-            });
-
-            confirmDelete.addEventListener('click', function() {
-                const articleId = deleteButton.getAttribute('data-id');
-                window.location.href = '/delete-article.php?id=' + articleId;
-            });
-
-            cancelDelete.addEventListener('click', function() {
-                deleteModal.style.display = 'none';
-            });
-        });
-    </script>
 </body>
